Improve compatibility

+++ includes/common.sh
@@ -411,18 +411,18 @@ _parse_webpage_and_get_url()
   if test ! -e "${MAIN_DIR:?}/cache/temp/headers"; then mkdir -p "${MAIN_DIR:?}/cache/temp/headers" || return "${?}"; fi
 
   {
-    _parsed_code="$("${WGET_CMD:?}" -q -S -O '-' "${@}" -- "${_url:?}")" 2> "${_headers_file:?}" || _status="${?}"
+    _parsed_code="$("${WGET_CMD:?}" -q -S -O '-' "${@}" -- "${_url:?}" 2> "${_headers_file:?}")" || _status="${?}"
 
     # shellcheck disable=SC3040 # Ignore: In POSIX sh, set option pipefail is undefined
     {
-      # IMPORTANT: We have to avoid "write error: Broken pipe" when a string is piped to "grep -q" or "grep -m 1"
+      # IMPORTANT: We have to avoid "printf: write error: Broken pipe" when a string is piped to "grep -q" or "grep -m 1"
       test "${USING_PIPEFAIL:-false}" = 'false' || set +o pipefail
-      _parsed_code="$(printf '%s\n' "${_parsed_code?}" | grep -o -m 1 -e "${_search_pattern:?}")" || _status="${?}"
+      _parsed_code="$(printf 2> /dev/null '%s\n' "${_parsed_code?}" | grep -o -m 1 -e "${_search_pattern:?}")" || _status="${?}"
       test "${USING_PIPEFAIL:-false}" = 'false' || set -o pipefail
     }
 
     if test "${_status:?}" -eq 0; then
-      test -n "${_parsed_code?}" || return 3
+      test -n "${_parsed_code?}" || return 10
       _parsed_url="$(printf '%s\n' "${_parsed_code:?}" | grep -o -e 'href=".*' | cut -d '"' -f '2' -s | sed -e 's|&amp;|\&|g')" || _status="${?}"
       if test "${DL_DEBUG:?}" = 'true'; then
         ui_debug "Parsed url: ${_parsed_url?}"
@@ -434,13 +434,14 @@ _parse_webpage_and_get_url()
       if test "${DL_DEBUG:?}" = 'true'; then
         ui_error_msg "Webpage parsing failed, error code => ${_status?}"
       fi
-      return 1
+      return 11
     fi
   }
 
+  test -s "${_headers_file:?}" || rm -f "${_headers_file:?}" || return "${?}" # Delete if empty
   _parse_and_store_all_cookies "${_domain:?}" 0< "${_headers_file:?}" || {
     ui_error_msg "Header parsing failed, error code => ${?}"
-    return 2
+    return 12
   }
   rm -f "${_headers_file:?}" || return "${?}"
